requirements.txt
dotenv==0.9.9
pandas==2.3.3
pandera==0.28.1
psycopg-binary==3.3.2
pytest==9.0.2
pytest-cov==7.0.0
altair==6.0.0
supabase==2.27.2